#451: Move share button to common
