+++ CourageScores/ClientApp/src/components/division_fixtures/scores/GameDetails.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {ShareButton} from "../../ShareButton";
+import {ShareButton} from "../../common/ShareButton";
 import {valueChanged} from "../../../helpers/events";
 
 export function GameDetails({ saving, access, fixtureData, setFixtureData }) {
