+++ CourageScores/ClientApp/src/components/division_teams/TeamOverview.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Link} from "react-router-dom";
 import {DivisionPlayers} from "../division_players/DivisionPlayers";
-import {ShareButton} from "../ShareButton";
+import {ShareButton} from "../common/ShareButton";
 import {any} from "../../helpers/collections";
 import {renderDate} from "../../helpers/rendering";
 import {useDivisionData} from "../DivisionDataContainer";
